refactor(lerobot-dataset): make decode helpers importable without the teleop extra
Move dataset_row_to_sim_qpos, sim_qpos_to_dataset_row and SO101_GRIPPER_LIMITS_RAD
into a new lerobot-free module so101_nexus.lerobot_dataset and export them from
the top-level so101_nexus package. They previously lived in
so101_nexus.lerobot_adapter.normalization, whose module hard-imports lerobot, so
a training or dataset consumer without the teleop extra (e.g. an RL repo) could
not import them at all. The helpers depend only on numpy and the joint-name
config.

normalization.py now imports the shared GripperLimitsRad alias and gripper-limit
validation from the new module, and the lerobot_adapter package keeps re-exporting
the helpers for backward compatibility. Add a test that the decode path imports
with lerobot absent, consolidate the pure-decode tests into the new lerobot-free
test module, and update the docs example to the top-level import.
