+++ src/so101_nexus/lerobot_adapter/__init__.py
@@ -7,16 +7,16 @@ runs LeRobot ChoiceRegistry decorators so ``--robot.type=sim_so_follower`` and
 
 from __future__ import annotations
 
-from so101_nexus.lerobot_adapter.normalization import (
-    SO101_GRIPPER_LIMITS_RAD,
-    dataset_row_to_sim_qpos,
-    read_gripper_limits_rad,
-    sim_qpos_to_dataset_row,
-)
+from so101_nexus.lerobot_adapter.normalization import read_gripper_limits_rad
 from so101_nexus.lerobot_adapter.sim_camera import SimCamera
 from so101_nexus.lerobot_adapter.sim_camera_config import SimCameraConfig
 from so101_nexus.lerobot_adapter.sim_follower import SimSOFollower
 from so101_nexus.lerobot_adapter.sim_follower_config import SimSOFollowerConfig
+from so101_nexus.lerobot_dataset import (
+    SO101_GRIPPER_LIMITS_RAD,
+    dataset_row_to_sim_qpos,
+    sim_qpos_to_dataset_row,
+)
 
 __all__ = [
     "SO101_GRIPPER_LIMITS_RAD",
